+++ CHANGELOG.md
@@ -2,6 +2,13 @@
 All notable changes to this project will be documented in this file.
 This project adheres to [Semantic Versioning](http://semver.org/).
 
+## [Unreleased]
+### Changed
+- Structure of files to more easily add new endpoints
+
+### Added
+- `rates` endpoint
+
 ## [0.4.0] - 2016-09-01
 ### Added
 - Made response `#success` a virtus attribute, so that it can be set in factories