+++ CHANGELOG.md
@@ -2,7 +2,7 @@
 All notable changes to this project will be documented in this file.
 This project adheres to [Semantic Versioning](http://semver.org/).
 
-## [Unreleased]
+## [0.7.0] - 2016-09-16
 ### Changed
 - Heavy lifting of sandbox and host defaults fall on BloomRemitClient::Client
 - Default sandbox is true. Must set to false to be on production mode